requirements.txt
Flask
Flask_SQLAlchemy
Flask_Bcrypt
Flask_JWT_Extended
python-dotenv
